double timeseries type

+++ HashFunctions/HashInterface.hpp
@@ -15,7 +15,7 @@ class HashInterface
     public:
         HashInterface(int w, int k, int TableSize, int vecSize, int probes);
         virtual ~HashInterface();
-        virtual unsigned int Hash(const std::vector<int> &p) = 0;
+        virtual unsigned int Hash(const std::vector<double> &p) = 0;
         inline int getNumProbes(){return numProbes;};
         virtual int *HashIndex(unsigned int ID, bool insertingMode = false) = 0;
 };
